requirements.txt
absl-py==2.1.0
action-msgs==1.2.1
actionlib-msgs==4.2.4
addict==2.4.0
ament-clang-format==0.12.11
ament-cmake-test==1.3.11
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
asttokens==3.0.0
attrs==24.3.0
blinker==1.9.0
builtin-interfaces==1.2.1
camera-calibration==3.0.6
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
cmake==3.31.2
comm==0.2.2
composition-interfaces==1.2.1
ConfigArgParse==1.7
contourpy==1.2.1
control-msgs==4.7.0
controller-manager==2.44.0
controller-manager-msgs==2.44.0
customtkinter==5.2.2
cv-bridge==3.2.1
cycler==0.12.1
darkdetect==0.8.0
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
diagnostic-msgs==4.2.4
diagnostic-updater==4.0.0
domain-coordinator==0.10.0
etils==1.11.0
evdev==1.7.1
example-interfaces==0.9.3
exceptiongroup==1.2.2
executing==2.1.0
Farama-Notifications==0.0.4
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.12.0
gazebo-msgs==3.7.0
generate-parameter-library-py==0.3.8
geometry-msgs==4.2.4
glfw==2.8.0
graph-msgs==0.2.0
gymnasium==0.28.1
h5py==3.12.1
idna==3.10
image-geometry==3.2.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
interactive-markers==2.3.2
ipython==8.31.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jax-jumpy==1.0.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
laser-geometry==2.4.0
launch==1.0.7
launch-pal==0.1.15
launch-param-builder==0.1.1
launch-pytest==1.0.7
launch-ros==0.19.8
launch-system-modes==0.7.1
launch-testing==1.0.7
launch-testing-examples==0.15.3
launch-testing-ros==0.19.8
launch-xml==1.0.7
launch-yaml==1.0.7
lifecycle-msgs==1.2.1
lit==18.1.8
llvmlite==0.43.0
map-msgs==2.1.0
MarkupSafe==3.0.2
matplotlib==3.7.2
matplotlib-inline==0.1.7
message-filters==4.3.5
moveit-configs-utils==2.5.6
moveit-msgs==2.2.1
mpmath==1.3.0
mujoco==3.2.6
nav-msgs==4.2.4
nbformat==5.7.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
open3d==0.17.0
opencv-python==4.8.0.74
openvr==1.26.701
osrf-pycommon==2.1.4
packaging==24.1
pandas==2.2.3
parso==0.8.4
pcl-msgs==1.0.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pyglet==1.5.27
Pygments==2.19.1
pynput==1.7.7
PyOpenGL==3.1.7
pyparsing==3.0.9
pyquaternion==0.9.9
pyrealsense2==2.54.1.5217
pyserial==3.4
python-dateutil==2.9.0.post0
python-qt-binding==1.1.2
python-xlib==0.33
pytz==2024.2
PyYAML==6.0.2
qt-gui==2.2.3
qt-gui-cpp==2.2.3
rcl-interfaces==1.2.1
rclpy==3.3.15
rcutils==5.1.6
realsense2-camera-msgs==4.55.1
referencing==0.35.1
requests==2.32.3
resource-retriever==3.1.2
retrying==1.3.4
rmw-dds-common==1.6.0
-e git+https://github.com/dwaitbhatt/robosuite.git@065a53d0389f3fe654871ba69f2edebe6e3fcbf7#egg=robosuite
ros2bag==0.15.13
ros2cli==0.18.11
ros2controlcli==2.44.0
ros2launch==0.19.8
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2test==0.4.0
ros2topic==0.18.11
rosbag2-interfaces==0.15.13
rosbag2-py==0.15.13
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpds-py==0.22.3
rpyutils==0.2.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-joint-trajectory-controller==2.38.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.0
scipy==1.11.1
scripts==3.7.0
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
shape-msgs==4.2.4
six==1.16.0
srdfdom==2.0.5
stable-baselines3==2.0.0
stack-data==0.6.3
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
sympy==1.13.3
system-modes-msgs==0.9.0
tenacity==9.0.0
termcolor==2.5.0
tf2-geometry-msgs==0.25.9
tf2-kdl==0.25.9
tf2-msgs==0.25.9
tf2-py==0.25.9
tf2-ros-py==0.25.9
threadpoolctl==3.5.0
topic-tools==1.1.1
topic-tools-interfaces==1.1.1
torch==2.0.1
tqdm==4.67.1
traitlets==5.14.3
trajectory-msgs==4.2.4
transforms3d==0.4.1
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
urllib3==2.3.0
visualization-msgs==4.2.4
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
xacro==2.0.8
-e git+https://github.com/xArm-Developer/xArm-Python-SDK.git@536a9c74cbf837bea3a93c91175580702cf33df8#egg=xArm_Python_SDK
zipp==3.21.0
